# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DG Toolkit API",
    "version": "1.0.0",
    "description": "KPMG Tunisia Data Governance Maturity Assessment Toolkit"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api"
    }
  ],
  "paths": {
...
    "/assessments": {
      "get": {
        "summary": "List all assessments",
        "tags": ["Assessments"],
        "responses": {
          "200": { "description": "List of assessments" }
        }
      },
      "post": {
        "summary": "Create a new assessment",
        "tags": ["Assessments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["organization_id", "consultant_id"],
                "properties": {
                  "organization_id": { "type": "integer" },
                  "consultant_id":   { "type": "integer" },
                  "engagement_date": { "type": "string", "format": "date" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Assessment created" },
          "404": { "description": "Organization or consultant not found" }
        }
      }
    },
    "/assessments/consultant/{consultant_id}": {
      "get": {
        "summary": "Get all assessments by consultant",
        "tags": ["Assessments"],
        "parameters": [
          { "name": "consultant_id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": {
          "200": { "description": "Assessments returned successfully" },
          "404": { "description": "Consultant not found" }
        }
      }
    },
    "/assessments/{assessment_id}": {
      "get": {
        "summary": "Get assessment by ID",
        "tags": ["Assessments"],
        "parameters": [
          { "name": "assessment_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Assessment found" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Soft delete an assessment",
        "tags": ["Assessments"],
        "parameters": [
          { "name": "assessment_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Assessment deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/assessments/{assessment_id}/targets": {
      "post": {
        "summary": "Set domain targets for an assessment",
        "tags": ["Assessments"],
        "parameters": [
          { "name": "assessment_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["targets"],
                "properties": {
                  "targets": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "domain_id":    { "type": "integer" },
                        "target_level": { "type": "integer", "minimum": 1, "maximum": 5 }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Targets set" },
          "409": { "description": "Targets already locked" }
        }
      }
    },
    "/assessments/{assessment_id}/start": {
      "post": {
        "summary": "Lock targets and start assessment",
        "tags": ["Assessments"],
        "parameters": [
          { "name": "assessment_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Assessment started" },
          "400": { "description": "Not all targets set" },
          "409": { "description": "Already started" }
        }
      }
    },
...
  }
}